src/test/kotlin/com/ginsberg/advent2019/Day17Test.kt
/*
 * Copyright (c) 2019 by Todd Ginsberg
 */

package com.ginsberg.advent2019

import com.ginsberg.advent2019.Resources.resourceAsString
import kotlinx.coroutines.FlowPreview
import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test

@DisplayName("Day 17")
@FlowPreview
class Day17Test {

    @Nested
    @DisplayName("Part 1")
    inner class Part1 {

        @Test
        fun `Actual answer`() {
            // Given
            val input = resourceAsString("day17.txt")

            // When
            val answer = Day17(input).solvePart1()

            // Then
            assertThat(answer).isEqualTo(11_372)
        }
    }

    @Nested
    @DisplayName("Part 2")
    inner class Part2 {
        @Test
        fun `Actual answer`() {
            // Given
            val input = resourceAsString("day17.txt")

            // When
            val answer = Day17(input).solvePart2()

            // Then
            assertThat(answer).isEqualTo(1_155_497)
        }
    }
}